cm-render: wire render_widget into screens_faithful + visual diff harness
Rewire the three widget helpers in `screens_faithful.rs` to construct
Widget records and delegate to `render_widget` (the 605-instruction
byte-exact FUN_005d7aa0 port), removing their hand-rolled primitive
call sequences.

- sidebar_button: was calling draw_panel(style=0x1021) then re-emitting
  8 draw_line calls for the 3D bevel. draw_panel already emits the
  P_BEVEL edges internally, so the old helper was double-drawing them.
  Now: builds a Widget with style_byte=0x1061 (0x1021 | 0x40 for the
  2-px label indent captured in the fixture) and calls render_widget.
- menu_item: was hand-drawing a two-line engraved separator with
  hardcoded colours 0x0180/0x0300. draw_panel's P_MIDLINE_H bit does
  the same thing with colours scaled from the row band colour, which
  is what the exe actually does. Now: builds a Widget with
  style_byte=0x10 (row) or 0x1000010 (separator) and delegates.
- menu_dropdown_container: unchanged behaviour, now routed through
  render_widget with flags |= 0x400 to suppress the empty label pass.

Widget.style_byte widened from u8 to u32 to match the asm — the exe's
[ebp+0x38] is read as a dword via `mov eax, [ebp+0x38]`, and the
captured widget-spec has styles that don't fit in a byte
(0x1021, 0x1000010). Low-byte bit tests still work on the low byte
of the dword.

Tests updated:
- All 5 existing screens_faithful tests kept, three of them
  restated to check what draw_panel actually produces (bevel
  non-zero, midline dark != bright with both non-zero) rather than
  what the old hand-drawn code did (hardcoded 0x0180/0x0300).
- New test file crates/cm-render/tests/visual_diff_sidebar_and_menu.rs:
    * `helpers_paint_something_in_the_expected_region` runs
      unconditionally — verifies the manager-menu fill and sidebar-
      button bevel land in the expected rects.
    * `diff_against_captured_exe_framebuffer` is #[ignore]d until
      fixtures/visual_sidebar_and_menu.pixels.bin lands (produced by
      tools/gdi_capture/capture_screen.py — the module docs carry the
      exact command). When the fixture exists the test does a byte-
      exact pixel-by-pixel comparison and, on mismatch, prints the
      diff count and the first 10 (x, y, expected, got) triples.

No bugs surfaced by the diff harness (the exe-comparison test is
gated on a fixture that must be produced against a live cm0102_GDI.exe
run — not something this commit can generate). The eliminated
double-bevel and hardcoded-separator-colour paths were real
approximations vs. the fixture spec; every widget primitive is now
sourced from render_widget's byte-exact dispatch.

cm-render lib tests: 416 pass (1 pre-existing failure in
packed_glyph::draw_text_walks_bytes_and_substitutes_pipe_for_space,
unrelated — reproduces on HEAD before this commit).

+++ crates/cm-render/src/packed_widget.rs
@@ -82,7 +82,14 @@ pub struct Widget {
     /// * `0x10` "no background restore" — suppresses block C save_rect
     /// * `0x20` "hover state" — swaps colour to hovered variant
     /// * `0x40` "pressed" — indents label by (2, 2) via `esi = ebx = 2`
-    pub style_byte: u8,
+    ///
+    /// Widened to `u32` on 2026-09-03 to match the asm — `mov eax,
+    /// [ebp+0x38]` reads a dword, and callers legitimately set high
+    /// bits (e.g. `0x1000` = P_SAMPLE_BG on sidebar buttons, `0x1000010`
+    /// = P_MIDLINE_H | P_SOLID_FILL on menu separators). The low-byte
+    /// bit tests (`al & 0x20`, `al & 0x40`) work on the low byte of the
+    /// dword regardless.
+    pub style_byte: u32,
 
     /// `+0x3c` — passed through to the label's `draw_wrapped_text` as
     /// the 5th positional arg (the `style` word from packed_text).
@@ -624,7 +631,7 @@ pub fn render_widget(
     // 005d7bff  mov cx, [ebp+0x72]        ; cx = colour_a
     // 005d7c03  xor esi, esi              ; label y offset = 0
     // 005d7c05  xor ebx, ebx              ; label x offset = 0
-    let mut effective_style: u32 = widget.style_byte as u32;
+    let mut effective_style: u32 = widget.style_byte;
     let mut panel_colour: u16 = widget.colour_a;
     let mut label_offset_x: i32 = 0;
     let mut label_offset_y: i32 = 0;
